backend/requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
Django==4.1.1
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
environs==9.5.0
marshmallow==3.17.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
phonenumberslite==8.12.55
Pillow==9.2.0
platformdirs==2.5.2
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
